user-service/requirements.txt
fastapi
uvicorn
databases[postgresql]
asyncpg
psycopg2-binary
bcrypt
python-jose[cryptography]
pydantic[email]
httpx
